+++ src/block/block_slvg.c
@@ -103,7 +103,8 @@ __wt_block_salvage_next(WT_SESSION_IMPL *session,
 		 * and get a page length from it.  Move to the next allocation
 		 * sized boundary, we'll never consider this one again.
 		 */
-		WT_ERR(__wt_read(session, fh, offset, allocsize, tmp->mem));
+		WT_ERR(__wt_read(
+		    session, fh, offset, (size_t)allocsize, tmp->mem));
 		blk = WT_BLOCK_HEADER_REF(tmp->mem);
 		block->slvg_off += allocsize;
